requirements.txt
streamlit
tensorflow
pillow>=10.0.0
numpy
